fix(nofork-reception): invalid argument.

+++ src/Kitchen/Cook.hpp
@@ -13,10 +13,11 @@
 #include "../Pizza.hpp"
 #include "../Wrappers/BalancingSemaphore.hpp"
 
-class Cook {
+class Cook
+{
 private:
     const uint16_t _id;
-    float _cookingTime;
+    double _cookingTime;
     Pizza _pizza;
 
     bool _isBusy = false;
@@ -24,20 +25,22 @@ private:
     std::chrono::system_clock::time_point _endCookingTime;
 
 public:
-    Cook(const uint16_t id, const float cookingTime) : _id(id), _cookingTime(cookingTime) {}
+    Cook(const uint16_t id, const double cookingTime) : _id(id), _cookingTime(cookingTime)
+    {
+    }
 
-    void cook(Pizza pizza, const std::vector<std::shared_ptr<BalancingSemaphore>> &ingredients, std::mutex &mutex,
-              std::ostream &os);
+    void cook(Pizza pizza, const std::vector<std::shared_ptr<BalancingSemaphore>>& ingredients, std::mutex& mutex,
+              std::ostream& os);
 
     [[nodiscard]] uint16_t getId() const { return _id; }
-    [[nodiscard]] float getCookingTime() const { return _cookingTime; }
+    [[nodiscard]] double getCookingTime() const { return _cookingTime; }
     [[nodiscard]] Pizza getPizza() const { return _pizza; }
     [[nodiscard]] bool isBusy() const { return _isBusy; }
     [[nodiscard]] std::chrono::system_clock::time_point getEndCookingTime() const { return _endCookingTime; }
 
     void setBusy(const bool isBusy) { _isBusy = isBusy; }
 
-    friend std::ostream &operator<<(std::ostream &os, const Cook &cook);
+    friend std::ostream& operator<<(std::ostream& os, const Cook& cook);
 };
 
 #endif // COOK_HPP
